Move testing to Docker and add documentation
Because os-config makes calls to systemd and mocking that is difficult,
not to mention not every host has systemd installed, tests are moved to
run in a Docker container with systemd installed and modified to work
with the containerized runtime.

This removes systemd as a factor that can cause tests to fail. Tests should
be run as documented in the Testing section of the README, as `cargo test`
will fail and also create some mock services on your dev machine that need
to be manually removed if the tests don't succeed.

Change-type: patch
Signed-off-by: Christina Ying Wang <christina@balena.io>