Modifying `config_json` in-place inside `handle_override_directives`
* Renamed `generate_config_json_migration` to `migrate_config_json`

* Introduced a `OverridesMap` type alias for `HashMap<String, serde_json::Value>`

* Renamed a `to_update` function argument to `overrides` to match the remote
config instructions name.

* Renamed `handle_update_directives` to `handle_override_directives` in
a similar fashion.

* `handle_override_directives` modifies `config_json` in-place instead
of operating on top of a clone. It returns a `bool` now to indicate
whether modifications were made. All this simplifies a bit the external
logic as well.

* Processing of `overrides` keys is now done in sorted order, so that
logging always produces the same output when the arguments of
`handle_override_directives` are the same. This helps with test code as
some of it operates on top of logging output.

* Inside the `migrate_config_json` integration tests I switched the output
checks to use the `.stdout(output)` pattern used in other tests as it
provides a diff output when something fails and is easier to navigate when
error occurs.

Signed-off-by: Zahari Petkov <zahari@balena.io>